requirements.txt
# HTTP and Web
requests>=2.31.0

# RSS Parsing
feedparser>=6.0.10

# Configuration Management
PyYAML>=6.0.1

# Natural Language Processing & Machine Learning
scikit-learn>=1.3.0

# LLM Integration
google-genai>=0.3.0